requirements.txt
duckduckgo-search>=4.1.1
langchain>=0.1.0
langchain-core>=0.1.10
pydantic>=2.5.0
requests>=2.31.0
readability-lxml>=0.8.1
html2text>=2020.1.16
lxml[html_clean]>=5.1.0
openai>=1.12.0
langchain-openai>=0.0.5
langgraph>=0.0.26
ipython>=8.22.1
jupyter>=1.0.0
pygraphviz>=1.12
graphviz>=0.20.1 